salt: add a precheck in downgrade to ensure its possible
We now set a boolean in `defautls.yaml` telling if we
support downgrade from a version or not.
It should be set to false if manual operations are needed
(e.g. downgrade of etcd), prior to downgrading the cluster
The cluster can still be force setting `force_downgrade`
to `True` in the pillar.
The check will fail early before anything got upgraded
to avoid having an half upgraded cluster situation.

Refs: #2852