scripts: add a --force flag to downgrade script
This option is used to by-pass the precheck in
downgrade orchestrate, preventing a user from
downgrading for unsupported path.
This is here in case one still want to downgrade
its cluster, for example, if he already has done
some manual operations to ensure the downgrade
will work (e.g. manual downgrade of etcd).

Refs: #2852